requirements.txt
kagglehub

langgraph
langchain-text-splitters
langchain-community
faiss-cpu
langchain_gigachat


rank_bm25
pickle